media: atomisp: sh_css_internal: Unifying sh_css_sp_group to remove #ifdef ISP2401
Since some parts of the data scructure elements are determined in compile
time, the configuration data structure should be compiled for both two
ISP models. In order to set the configuration for both ISP model in
runtime, Thesh_css_sp_group is unified to one data structure for the
configuration to ensure the data structure can be used for both ISP2400
and 2401.

Signed-off-by: Kate Hsuan <hpa@redhat.com>