media: stomisp: make two enum (mipi_format_t) and remove unused code
The original mipi_format_t represents the image format settings for both
ISP model. However, if we want to make this driver generic, this enum
should be individialy defined for them. Moreover, the function was not used
by any functions so it was removed.

Signed-off-by: Kate Hsuan <hpa@redhat.com>

staging: atomisp: rename the MIPI format for both type of atomisp

Rename the enum for both types of atomisp.

Signed-off-by: Kate Hsuan <hpa@redhat.com>

staging: atomisp: redefine the format for both model

There is a variable redefinition issue about the video format if #ifdef
ISP2401 is removed. Therefore, the video format needs to be redefined for
both quirks of ISP.

Revert somting. Need put this patch to the lastest of the series.