leds: rgb: leds-ktd202x: Get device properties through fwnode to support ACPI
This LED controller also installed on a Xiaomi pad2 and it is a x86
platform. The original driver is based on device tree and can't be
used for this ACPI based system. This patch migrated the driver to
use fwnode to access the properties. Moreover, the fwnode API
supports device tree so this work would effect the original
implementations.

Signed-off-by: Kate Hsuan <hpa@redhat.com>