media: atomisp: t4ka3: Switch to using standard v4l2_ctrl_new_foo() helpers
Rework the v4l2-ctrl handling to follow standard best practices using
standard v4l2_ctrl_new_foo() helpers insytead of using
v4l2_ctrl_new_custom() for all controls.

This also involves changing the V4L2_CID_TEST_PATTERN and
V4L2_CID_LINK_FREQ controls to menus to match the v4l2 API specification
which says that these controls must be menu controls.

This is a preparation patch for not powering on and thus not touching
the sensor when it is not streaming, letting the v4l2-ctrl framework cache
the control values and then call __v4l2_ctrl_handler_setup() on stream on
(after powering on the sensor) to set their values before sending
the stream on command to the sensor.

Only powering on the sensor itself is a preparation for adding runtime-pm
support, which in turn is preparation for switching from custom atomisp-gmin
power-sequencing (which does not work with the Whiskey Cove PMIC) to using
ACPI for powering on/off the sensor.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>