media: atomisp: t4ka3: switch to ACPI power manegement
The power of the sensor, including clk and regulator are controlled by
the APCI power management and they can be found in XiaoMi Pad2 DSDT
information.

Since the power management is not through the proper atomisp_gmin,
atomisp_register_sensor_no_gmin() / atomisp_unregister_subdev() are used
to register/unregister the sensor and keep the ACPI power management.

It has been tested on XiaoMi Pad2

Signed-off-by: Kate Hsuan <hpa@redhat.com>